feat(tsconfig): move resolveTsconfigPaths function to a new module for better organization
